refactor: consolidate mongo client options integration tests with other client tests
