+++ test/integration/node-specific/mongo_client.test.js
@@ -1,12 +1,15 @@
 'use strict';
 const { expect } = require('chai');
 
+const sinon = require('sinon');
+
 const { setupDatabase, assert: test } = require('../shared');
 const { format: f } = require('util');
 
-const { ReadPreference } = require('../../../src');
+const { MongoClient, ReadPreference } = require('../../../src');
 const { Db } = require('../../../src/db');
-const { getTopology } = require('../../../src/utils');
+const { Connection } = require('../../../src/cmap/connection');
+const { getTopology, isHello } = require('../../../src/utils');
 
 describe('MongoClient integration', function () {
   before(function () {
@@ -233,5 +236,96 @@ describe('MongoClient integration', function () {
     const client = this.configuration.newClient({}, { readPreference: ReadPreference.SECONDARY });
     expect(client.readPreference).to.be.instanceOf(ReadPreference);
     expect(client.readPreference).to.have.property('mode', ReadPreference.SECONDARY);
+  });
+
+  it('should error on unexpected options', {
+    metadata: { requires: { topology: 'single' } },
+
+    test: function (done) {
+      var configuration = this.configuration;
+      MongoClient.connect(
+        configuration.url(),
+        {
+          maxPoolSize: 4,
+          notlegal: {},
+          validateOptions: true
+        },
+        function (err, client) {
+          expect(err)
+            .property('message')
+            .to.match(/options notlegal, validateoptions are not supported/);
+          expect(client).to.not.exist;
+          done();
+        }
+      );
+    }
+  });
+
+  it('should error on unexpected options (promise)', {
+    metadata: { requires: { topology: 'single' } },
+
+    test() {
+      MongoClient.connect(this.configuration.url(), {
+        maxPoolSize: 4,
+        notlegal: {},
+        validateOptions: true
+      })
+        .then(() => expect().fail())
+        .catch(err => {
+          expect(err)
+            .property('message')
+            .to.match(/options notlegal, validateoptions are not supported/);
+        });
+    }
+  });
+
+  it('must respect an infinite connectTimeoutMS for the streaming protocol', {
+    metadata: { requires: { topology: 'replicaset', mongodb: '>= 4.4' } },
+    test: function (done) {
+      const client = this.configuration.newClient({
+        connectTimeoutMS: 0,
+        heartbeatFrequencyMS: 500
+      });
+      client.connect(err => {
+        expect(err).to.not.exist;
+        const stub = sinon.stub(Connection.prototype, 'command').callsFake(function () {
+          const args = Array.prototype.slice.call(arguments);
+          const ns = args[0];
+          const command = args[1];
+          const options = args[2] || {};
+          if (ns.toString() === 'admin.$cmd' && isHello(command) && options.exhaustAllowed) {
+            expect(options).property('socketTimeoutMS').to.equal(0);
+            stub.restore();
+            client.close(done);
+          }
+          stub.wrappedMethod.apply(this, args);
+        });
+      });
+    }
+  });
+
+  it('must respect a finite connectTimeoutMS for the streaming protocol', {
+    metadata: { requires: { topology: 'replicaset', mongodb: '>= 4.4' } },
+    test: function (done) {
+      const client = this.configuration.newClient({
+        connectTimeoutMS: 10,
+        heartbeatFrequencyMS: 500
+      });
+      client.connect(err => {
+        expect(err).to.not.exist;
+        const stub = sinon.stub(Connection.prototype, 'command').callsFake(function () {
+          const args = Array.prototype.slice.call(arguments);
+          const ns = args[0];
+          const command = args[1];
+          const options = args[2] || {};
+          if (ns.toString() === 'admin.$cmd' && isHello(command) && options.exhaustAllowed) {
+            expect(options).property('socketTimeoutMS').to.equal(510);
+            stub.restore();
+            client.close(done);
+          }
+          stub.wrappedMethod.apply(this, args);
+        });
+      });
+    }
   });
 });
